// frontend/app/types/profile.ts
export interface UserProfile {
    firstName: string;
    lastName: string;
    country: string;
    address: string;
    phoneNumber?: string;
    education: string;
    workExperience: string[];
    skills: string[];
    certificates: string[];
    languages: string[];
    additionalInformation: string;
    aboutMe: string;
  }
  